Move modeling utils to modeling
